Refactor: Marco the printing the results of calling functions.
